// src/types/bindings.ts
export type Anime = {
	id: number,
	judul: string,
	japanese: string | null,
	produser: string | null,
	tanggal_rilis: string | null,
	studio: string | null,
	image_file: string,
	folder_name: string | null,
	genres: Genre[],
};
// ...
export type Genre = {
	id: number,
	name: string,
};

export type GenreCount = {
	name: string,
	total: number,
};

export type GenreInformation = {
	name: string,
	display: string,
};

export type InitStatus = {
	is_initialized: boolean,
	has_store_anime_directory: boolean,
	has_youtube_download_directory: boolean,
};

export type LibraryAnimeInformation = {
	anime: Anime,
	episodes: LibraryEpisodeInformation[],
};

export type LibraryEpisodeInformation = {
	name: string,
	resolutions: Resolution[],
};

export type Mirror = {
	name: string,
	data: MirrorData,
};

export type MirrorData = {
	id: number,
	i: number,
	q: string,
};

export type MirrorLink = {
	url: string,
	nonce: string,
};

export type Platform = "Pdrain" | "Mega" | "DesuDrive";

export type RenameMapping = {
	original_path: string,
	new_name: string,
};

export type Resolution = "P360" | "P480" | "P720";